fix: Actual slips

+++ onyx/onyx/app/window.cpp
@@ -12,10 +12,10 @@ namespace Onyx
 
 static TKit::BlockAllocator createAllocator()
 {
-    const u32 maxSize =
-        std::max({sizeof(RenderContext<D2>), sizeof(RenderContext<D3>), sizeof(Camera<D2>), sizeof(Camera<D3>)});
-    const u32 alignment =
-        std::max({alignof(RenderContext<D2>), alignof(RenderContext<D3>), alignof(Camera<D2>), alignof(Camera<D3>)});
+    const u32 maxSize = static_cast<u32>(
+        std::max({sizeof(RenderContext<D2>), sizeof(RenderContext<D3>), sizeof(Camera<D2>), sizeof(Camera<D3>)}));
+    const u32 alignment = static_cast<u32>(
+        std::max({alignof(RenderContext<D2>), alignof(RenderContext<D3>), alignof(Camera<D2>), alignof(Camera<D3>)}));
     return TKit::BlockAllocator{maxSize * 2 * (ONYX_MAX_RENDER_CONTEXTS + ONYX_MAX_CAMERAS), maxSize, alignment};
 }
